==> notebooks/Mozambique/5_Do_mosaic_clipping.ipynb <==
Notebook to mosaic tiles and clip to country boundary.

### Mosaic probabilities, predictions and post-processed tiles

! gdal_merge.py -o Results/Mozmabique_land_cover_prediction_mosaic.tif -co COMPRESS=Deflate -ot Byte Results/Mozmabique_land_cover_prediction_tile_*.tif -co BIGTIFF=YES

! gdal_merge.py -o Results/Mozmabique_land_cover_probability_mosaic.tif -co COMPRESS=Deflate -ot Byte Results/Mozmabique_land_cover_probability_tile_*.tif -co BIGTIFF=YES

! gdal_merge.py -o Results/Mozmabique_land_cover_prediction_postprocessed_mosaic.tif -co COMPRESS=Deflate -ot Byte Results/Mozmabique_land_cover_prediction_postprocessed_tile_*.tif -co BIGTIFF=YES

### clip mosaics to country boundary

import subprocess
bound_file="Data/Mozambique_boundary.shp"

mosaic="Results/Mozmabique_land_cover_prediction_mosaic.tif"
mosaic_clipped="Results/Mozmabique_land_cover_prediction_mosaic_clipped.tif"
gdal_cmd=["gdalwarp", "-of", "GTiff", '-cutline',bound_file,'-crop_to_cutline',
          '-co','COMPRESS=DEFLATE',mosaic,mosaic_clipped]
p1=subprocess.run(gdal_cmd)

mosaic="Results/Mozmabique_land_cover_probability_mosaic.tif"
mosaic_clipped="Results/Mozmabique_land_cover_probability_mosaic_clipped.tif"
gdal_cmd=["gdalwarp", "-of", "GTiff", '-cutline',bound_file,'-crop_to_cutline',
          '-co','COMPRESS=DEFLATE',mosaic,mosaic_clipped]
p1=subprocess.run(gdal_cmd)

mosaic="Results/Mozmabique_land_cover_prediction_postprocessed_mosaic.tif"
mosaic_clipped="Results/Mozmabique_land_cover_prediction_postprocessed_mosaic_clipped.tif"
gdal_cmd=["gdalwarp", "-of", "GTiff", '-cutline',bound_file,'-crop_to_cutline',
          '-co','COMPRESS=DEFLATE',mosaic,mosaic_clipped]
p1=subprocess.run(gdal_cmd)
